requirements.txt
pyodbc
python-dotenv
uvicorn[standard]
pydantic
azure-identity
json2xml
python-dateutil
pandas
matplotlib